feat!: removed RDListing and prefer direct rendering using flags
unroll struct fields and arrays in flags space
exhanced type system
added 'bool' type
calculate type size and field offsets at registration time
lots of fixes

+++ include/redasm/surface/renderer.h
@@ -9,7 +9,7 @@
 typedef struct RDRenderer RDRenderer;
 
 RD_API RDContext* rd_renderer_get_context(RDRenderer* self);
-RD_API void rd_renderer_ws(RDRenderer* self, int n);
+RD_API void rd_renderer_ws(RDRenderer* self, usize n);
 RD_API void rd_renderer_norm(RDRenderer* self, const char* s);
 RD_API void rd_renderer_text(RDRenderer* self, const char* s, RDThemeKind fg,
                              RDThemeKind bg);
